% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{do_bms}
\alias{do_bms}
\title{Perform BMS for classifying GxT interactions}
\usage{
do_bms(
  data,
  fn,
  rint = FALSE,
  method,
  p.m = NULL,
  phi,
  phi0 = sqrt(1000),
  kappa = 0.002,
  nu = 0.002,
  kappa.u = 0.002,
  nu.u = 0.002,
  tu.lambda = NULL,
  summary = TRUE,
  seed = 1,
  n.cores = 4
)
}
\arguments{
\item{data}{A list containing vectors of subjects, genotypes,
phenotypes, and treatment indicators, which must be named
"subject", "g", "y", and "t", respectively.}

\item{fn}{A character string specifying the function. This must be
one of "nonlinear" and "linear", corresponding to nonlinear and
linear models, respectively.}

\item{rint}{A logical as to whether the phenotypes need to be
RINT-transformed. This cannot be set to \code{TRUE} if \code{fn} is
set to "nonlinear".}

\item{method}{A character string spcifying the method for parameter
estimation and computation of the marginal likelihood. This
must be eigher "mcmc.bs", for MCMC followed by bridge sampling,
or "map.lap", for MAP estimation
followed by Laplace approximation.}

\item{p.m}{A vector of hyperparameters of the model prior.}

\item{phi}{A vector of hyperparameters of the effect prior.}

\item{phi0}{A scalar specifying the hyperparameter on the intercept.}

\item{kappa}{A scalar specifying the hyperparameter of the gamma
prior on the residual error precision.}

\item{nu}{A scalar specifying the hyperparameter of the gamma prior
on the residual error precision.}

\item{kappa.u}{A scalar specifying the hyperparameter of the gamma
prior on the random intercept.}

\item{nu.u}{A scalar specifying the hyperparameter of the gamma
prior on the random intercept.}

\item{tu.lambda}{A list obtained from
\code{\link{get_tu_lambda}}. It must contain the transposed
eigen vector matrix and the eigen values of the covariance
matrix. The element names must be "tU" and "lambda".}

\item{summary}{A Boolean variable indicating whether a summary or
MCMC samples should be stored. This is only applicable when
\code{method} is set to "mcmc.bs".}

\item{seed}{An integer specifying a seed for RNG.}

\item{n.cores}{An integer specifying the number of cores.}
}
\value{
A list object containing:
\itemize{
\item{\code{fn} -  A character string specifying the function.}
\item{\code{ranef} - A logical.}
\item{\code{rint} - A logical as to whether the phenotypes have
been RINT-transformed.}
\item{\code{p.m} - A vector of hyperparameters of the model prior.}
\item{\code{seed} - A integer specifying a seed fo RNG.}
\item{\code{ln.p.y.given.m} - A named vector of the log marginal
likelihood given each model. If \code{method} is set to
"map.lap", the value is shifted by a constant.}
\item{\code{ln.p.y} - A scalar value of the log marginal
likelihood. If \code{method} is set to "map.lap", the value is
shifted by a constant.}
\item{\code{p.m.given.y} - A named vector of posterior probability of the
models.}
\item{\code{ml.errors} - A named vector of errors in the log marginal
likelihood. This element is included only when \code{method} is
set to "mcmc.bs".}
\item{\code{stan.list} - A list of \code{stanfit} objects or data frames
containing MCMC results for all eight models. This element is inclued only when
\code{method} is set to "mcmc.bs". If \code{summary} is set to
\code{TRUE}, a list of data frames containing posterior summary
is returned. If the option is set to \code{FALSE}, a list of
\code{stanfit} objects is returned.}
\item{\code{optim.list} - A list of outputs from the \code{optim}
function from the \code{stat} package containing MAP
estimates and Hessian. This element is included only when
\code{method} is set to "map.lap".}
}
}
\description{
This function takes as input individual phenotype and genotype data
and performs BMS for a feature-SNP pair (or a SNP). The output
includes posterior probabilities of different types of GxT
interactions as well as the log marginal likelihood, which can be
used for optimizing the hyperparameter values by an empirical Bayes
approach. BMS is performed by either MCMC followed by bridge
sampling or MAP estimation followed by Laplace approximation. In
the latter case, a relative marginal likelihood is returned. That
is, the log marginal likelihood value is shifted by a
constant. Note that this shift does not affect the posterior
probabilities or hyperparameter optimization.
\code{tu.lambda} must be supplied when modeling a random effect.
}
