% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_input}
\alias{format_input}
\title{Format input}
\usage{
format_input(
  qtl,
  pheno,
  geno,
  covar,
  int,
  pheno.col = "phenotype_id",
  variant.col = "variant_id",
  feat.col = "gene_id",
  snp.col = "SNP",
  pheno.rm = 1:4,
  geno.rm = 1:6
)
}
\arguments{
\item{qtl}{A character string specfying the name of the output file
from interacton QTL mapping from TensorQTL.}

\item{pheno}{A character string specifynig the name of the file
containing molecular phenotes. This must be the same as the
input file used for TensorQTL.}

\item{geno}{A character string specifying the name of the genotype
file in \href{https://www.cog-genomics.org/plink/2.0}{PLINK2}
\code{.traw} format.}

\item{covar}{A character string specifyng the name of the file
containing covariates. This must be the same as the input file
used for TensorQTL.}

\item{int}{A character string specifying the name of the file
containing the condition variable of interest, coded as 1 and
2. This must be the same as the input file used for TensorQTL.}

\item{pheno.col}{A character string specifying the name of the
column of \code{qtl} containing the feature IDs.}

\item{variant.col}{A character string specifying the name of the
column of \code{qtl} containing the feature IDs.}

\item{feat.col}{A character string specifying the name of the
column of \code{pheno} containing the feature IDs.}

\item{snp.col}{A character string specifying the name of the
column of \code{geno} containing the SNP IDs.}

\item{pheno.rm}{A vector of integers corresponding to the column
numbers for non-phenotype entries in \code{pheno}.}

\item{geno.rm}{A vector of integers corresponding to the column
numbers for non-phenotype entries in \code{geno}.}
}
\value{
A list of lists containing:
\itemize{
\item{\code{y} - A vector of phenotypes.}
\item{\code{g} - A vector of genotypes.}
\item{\code{t} - A vector of treatment indicators.}
\item{\code{subject} - A vector of subject.}
\item{\code{feat.id} - A character string spcifyng the feature ID.}
\item{\code{snp.id} - A character string specifying the SNP ID.}
}
}
\description{
This function takes input and output files of
\href{https://github.com/broadinstitute/tensorqtl}{TensorQTL} and
returns a list object that can be used as input for
\code{\link{do_bms}}. It currenly only handles completely unpaired data,
where there is no repeated measurement per donor (subject). This
includes typical molecular data from clinical trials.
}
