% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_names}
\alias{get_model_names}
\title{Get the names of the eight models}
\usage{
get_model_names()
}
\value{
A vector containing the eight model names.
}
\description{
This function returns an ordered vector of character strings
corresponding to the names of the eight models.
}
